.github/scripts/requirements.txt
requests>=2.28.0
PyYAML>=6.0
configparser>=5.0.0
looker-sdk>=23.0.0
